Fix Modern-style Sounds Using Number Indices
PR #6112 protected retail sound indices from being overwritten with the intent of when modders used string names for sounds,. Though that PR ended up being a bit overprotective because it made it so that using the modern sound format but still using numbered names did not work. In other words, for mods that use the modern sound format but still use numbered names, FSO sees the new format but then sees the numbers are within a protected range and offsets them, which means the sound index values referenced in ships and weapons table now are not correct and thus do not play.

This PR fixes that edge-case. Tested and works as expected. Also, if folks would rather just rever the entire protective offset aspect I am fine with that and we can instead just make a warning to the modder that their sound name will overwrite a protected range (though the changes of this PR to check if the name is the index should still be included, since an overwrite warning is not needed if the index and name are already the same).