use proper stats block for display in barracks
There are 3 different stats blocks: two in the pilot file for single and
multi, and one in the campaign savefile. The stats block in the savefile
has historically been used in the barracks if possible since it's sanitized
for the current mod data. Unfortunately this means that the stats will change
if the campaign changes or is restarted.

However it is possible to easily produce a sanitized scoring struct based
on the stats blocks from the pilot file. This allows us to use true all-time
stats for display in the barracks while still accounting for mod differences.