add new variants of form-on-wing and stay-still
The original ai-form-on-wing and ai-stay-still goals automatically cleared all other goals assigned to the same ship or wing, which is not always desired.  This adds non-goal-clearing variants of these goals, similar to ai-play-dead and ai-play-dead-persistent.  There is also a game_settings flag to make the new variant of form-on-wing the standard for HUD squad messaging.

Also removes `AI_GOAL_PLACEHOLDER_1` since it is not needed; goals higher than `AI_GOAL_NONE` do not need to be a specific value.