Fix offset dialog option image with controls preset confirmation
The Popup Dialog for `Confirm Control Preset Overwrite` has the Yes and No buttons order switched compared to what they are usually set as. This means the images appear offset and do not look sharp against the background of the dialog box image. This PR switches the order to match and remove that visual bug. Albeit a very minor bug, but one I have heard from a few folks and noticed myself as well.

PR is tested and works as expected.