Allow Modders to Tune Guarding
Overall guarding behavior of ships can easily lead to collisions with other large ships if there is a convoy or group of large ships together. This PR adds the same logic of `combat collision avoidance` to also be enabled for guarding.  It also exposes the collision avoid aggression factors for both combat and guard collision aggression to allow modders to tune values to better improve collision avoidance for their mod.

This PR also exposes the radius at which the AI decides to use `big_guard` mode (ie orbit around a cylinder shape) or not. It also allows modders to have the AI slow down when guarding in `big_guard` mode.

In combination, this set of values allows modders to finely tune guarding behavior and more easily reduce collisions with large ships.

Everything is tested and works as expected.