Allow modders to always show first game tip at start
This PR adds a simple boolean that allows mods the ability to always show the game tips in the same order, starting with the first entry. This is particularly useful if the mod is introducing something to new players.

The current behavior is to pick a random starting location, so this PR just always starts at tip index 0.

Tested and works as expected.