Fix changes not properly discarding
Follow-up to #6595. I had thought I would have to make a backup function for every slider and value that was not getting properly discard, but it turns out the reason they were not getting properly discarded is b/c the rest of the on frame function was setting them after the escape button section.