Hide Disabled Controls in the F1 Screen
If a control is disabled it does not appear to players in the Controls Configure screen, thus it should also not appear to players in the F1 Gameplay Help screen.  This PR adds a simple early return within `gameplay_help_init_control_line` to fix this issue.

Tested and works as expected.