Allows mods to fully disable energy transfering
Mods such as FotG disable the controls for transferring energy between the shields and weapons for multiple reasons (such as being very difficult to balance, especially with varying weapon recharge rates). Even with the player having that ability disabled, the AI still currently is able to do this.

This PR creates a AI profiles flag to also prevent the AI from using the energy transfer ability. Note, it is set as a AI profiles flag since it's a higher order option related to if mods overall allow this ability or not.